requirements.txt
joblib==0.16.0
numpy==1.19.2
pandas==1.1.2
blosc==1.9.2
pillow==7.2.0
seaborn==0.11.0
scikit==learn==0.23.2